+++ internal/controller/admission/falconadmission_controller.go
@@ -13,6 +13,7 @@ import (
 	k8sutils "github.com/crowdstrike/falcon-operator/internal/controller/common"
 	"github.com/crowdstrike/falcon-operator/pkg/aws"
 	"github.com/crowdstrike/falcon-operator/pkg/common"
+	"github.com/crowdstrike/falcon-operator/pkg/falcon_secret"
 	"github.com/crowdstrike/falcon-operator/pkg/registry/pulltoken"
 	"github.com/crowdstrike/falcon-operator/pkg/tls"
 	"github.com/crowdstrike/falcon-operator/version"
@@ -235,6 +236,12 @@ func (r *FalconAdmissionReconciler) Reconcile(ctx context.Context, req ctrl.Requ
 		return ctrl.Result{}, err
 	}
 
+	if falconAdmission.Spec.FalconSecret.Enabled {
+		if err = r.injectFalconSecretData(ctx, falconAdmission, log); err != nil {
+			return ctrl.Result{}, err
+		}
+	}
+
 	configUpdated, err := r.reconcileConfigMap(ctx, req, log, falconAdmission)
 	if err != nil {
 		return ctrl.Result{}, err
@@ -377,14 +384,23 @@ func (r *FalconAdmissionReconciler) reconcileService(ctx context.Context, req ct
 	service := assets.Service(falconAdmission.Name, falconAdmission.Spec.InstallNamespace, common.FalconAdmissionController, selector, common.FalconAdmissionServiceHTTPSName, port)
 
 	err := common.GetNamespacedObject(ctx, r.Client, r.Reader, types.NamespacedName{Name: falconAdmission.Name, Namespace: falconAdmission.Spec.InstallNamespace}, existingService)
-	if err != nil && apierrors.IsNotFound(err) {
+
+	switch {
+	case err == nil && !falconAdmission.GetAdmissionControlEnabled():
+		err = k8sutils.Delete(r.Client, ctx, req, log, falconAdmission, &falconAdmission.Status, service)
+		if err != nil {
+			return false, err
+		}
+		return false, nil
+	case apierrors.IsNotFound(err) && falconAdmission.GetAdmissionControlEnabled():
 		err = k8sutils.Create(r.Client, r.Scheme, ctx, req, log, falconAdmission, &falconAdmission.Status, service)
 		if err != nil {
 			return false, err
 		}
-
 		return false, nil
-	} else if err != nil {
+	case apierrors.IsNotFound(err) && !falconAdmission.GetAdmissionControlEnabled():
+		return false, nil
+	case err != nil:
 		log.Error(err, "Failed to get FalconAdmission Service")
 		return false, err
 	}
@@ -404,7 +420,6 @@ func (r *FalconAdmissionReconciler) reconcileService(ctx context.Context, req ct
 func (r *FalconAdmissionReconciler) reconcileAdmissionValidatingWebHook(ctx context.Context, req ctrl.Request, log logr.Logger, falconAdmission *falconv1alpha1.FalconAdmission, cabundle []byte) (bool, error) {
 	existingWebhook := &arv1.ValidatingWebhookConfiguration{}
 	disabledNamespaces := append(common.DefaultDisabledNamespaces, falconAdmission.Spec.AdmissionConfig.DisabledNamespaces.Namespaces...)
-	const webhookName = "validating.admission.falcon.crowdstrike.com"
 	failPolicy := arv1.Ignore
 	port := int32(443)
 
@@ -431,18 +446,26 @@ func (r *FalconAdmissionReconciler) reconcileAdmissionValidatingWebHook(ctx cont
 		port = *falconAdmission.Spec.AdmissionConfig.Port
 	}
 
-	webhook := assets.ValidatingWebhook(falconAdmission.Name, falconAdmission.Spec.InstallNamespace, webhookName, cabundle, port, failPolicy, disabledNamespaces)
+	webhook := assets.ValidatingWebhook(falconAdmission.Name, falconAdmission.Spec.InstallNamespace, common.FalconAdmissionValidatingWebhookName, cabundle, port, failPolicy, disabledNamespaces)
 	updated := false
 
-	err = common.GetNamespacedObject(ctx, r.Client, r.Reader, types.NamespacedName{Name: webhookName}, existingWebhook)
-	if err != nil && apierrors.IsNotFound(err) {
+	err = common.GetNamespacedObject(ctx, r.Client, r.Reader, types.NamespacedName{Name: common.FalconAdmissionValidatingWebhookName}, existingWebhook)
+	switch {
+	case err == nil && !falconAdmission.GetAdmissionControlEnabled():
+		err = k8sutils.Delete(r.Client, ctx, req, log, falconAdmission, &falconAdmission.Status, webhook)
+		if err != nil {
+			return false, err
+		}
+		return false, nil
+	case apierrors.IsNotFound(err) && falconAdmission.GetAdmissionControlEnabled():
 		err = k8sutils.Create(r.Client, r.Scheme, ctx, req, log, falconAdmission, &falconAdmission.Status, webhook)
 		if err != nil {
 			return false, err
 		}
-
 		return false, nil
-	} else if err != nil {
+	case apierrors.IsNotFound(err) && !falconAdmission.GetAdmissionControlEnabled():
+		return false, nil
+	case err != nil:
 		log.Error(err, "Failed to get FalconAdmission Validating Webhook")
 		return false, err
 	}
@@ -594,7 +617,12 @@ func (r *FalconAdmissionReconciler) reconcileAdmissionDeployment(ctx context.Con
 }
 
 func (r *FalconAdmissionReconciler) reconcileRegistrySecret(ctx context.Context, req ctrl.Request, log logr.Logger, falconAdmission *falconv1alpha1.FalconAdmission) error {
-	pulltoken, err := pulltoken.CrowdStrike(ctx, r.falconApiConfig(ctx, falconAdmission))
+	apiConfig, err := r.falconApiConfig(ctx, falconAdmission)
+	if err != nil {
+		return err
+	}
+
+	pulltoken, err := pulltoken.CrowdStrike(ctx, apiConfig)
 	if err != nil {
 		return fmt.Errorf("unable to get registry pull token: %v", err)
 	}
@@ -705,6 +733,37 @@ func (r *FalconAdmissionReconciler) admissionDeploymentUpdate(ctx context.Contex
 	if err := k8sutils.Update(r.Client, ctx, req, log, falconAdmission, &falconAdmission.Status, existingDeployment); err != nil {
 		return err
 	}
+
+	return nil
+}
+
+func (r *FalconAdmissionReconciler) injectFalconSecretData(ctx context.Context, falconAdmission *falconv1alpha1.FalconAdmission, logger logr.Logger) error {
+	logger.Info("injecting Falcon secret data into Spec.Falcon and Spec.FalconAPI - sensitive manifest values will be overwritten with values in k8s secret")
+	falconSecret := &corev1.Secret{}
+	falconSecretNamespacedName := types.NamespacedName{
+		Name:      falconAdmission.Spec.FalconSecret.SecretName,
+		Namespace: falconAdmission.Spec.FalconSecret.Namespace,
+	}
+
+	if err := common.GetNamespacedObject(ctx, r.Client, r.Reader, falconSecretNamespacedName, falconSecret); err != nil {
+		return err
+	}
+
+	cid := falcon_secret.GetFalconCIDFromSecret(falconSecret)
+	falconAdmission.Spec.Falcon.CID = &cid
+
+	provisioningToken := falcon_secret.GetFalconProvisioningTokenFromSecret(falconSecret)
+	falconAdmission.Spec.Falcon.PToken = provisioningToken
+
+	if falconAdmission.Spec.FalconAPI == nil {
+		logger.Info("skipped injecting FalconAPI secrets - Spec.FalconAPI is nil")
+		return nil
+	}
+
+	clientId, clientSecret := falcon_secret.GetFalconCredsFromSecret(falconSecret)
+	falconAdmission.Spec.FalconAPI.ClientId = clientId
+	falconAdmission.Spec.FalconAPI.ClientSecret = clientSecret
+	falconAdmission.Spec.FalconAPI.CID = &cid
 
 	return nil
 }